+++ src/index.ts
@@ -1,0 +1,11 @@
+// Button Component
+export { default as Button } from './components/button'
+export type { ButtonProps } from './components/button'
+
+// Icon Component
+export { default as Icon } from './components/icon'
+export { iconNames } from './components/icon'
+export type { IconTypes } from './components/icon'
+
+// Tools
+export { cn } from './utils'