+++ src/utils.ts
@@ -1,0 +1,2 @@
+export const cn = (...args: (string | boolean | undefined)[]): string =>
+    args.filter((item: any) => !!item).join(' ')