+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # simple-react-ui-kit
 
+## 1.0.3
+
+### Patch Changes
+
+-   Rebuild UI architecture
+-   Fixed Button component styles
+-   Added story for Button component
+-   Rename concat class names functions (cn)
+
 ## 1.0.2
 
 ### Patch Changes